feat: higlight lines to code options and simplified (without color options as highlighting behavior changed)
